requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-dotenv==1.1.0
jinja2==3.1.6
alembic==1.15.2
pymupdf==1.25.5 